README
This project was completed by BAROUDI ISMAËL in partial fulfilment of ECON-UB.0232, Data Bootcamp, Spring 2018. I certify that the NYU Stern Honor Code applies to this project. In particular, I have:
Clearly acknowledged the work and efforts of others when submitting written work as our own. The incorporation of the work of others–including but not limited to their ideas, data, creative expression, and direct quotations (which should be designated with quotation marks), or para- phrasing thereof– has been fully and appropriately referenced using notations both in the text and the bibliography.

And I understand that:
Submitting the same or substantially similar work in multiple courses, either in the same semester or in a different semester, without the express approval of all instructors is strictly forbidden.
I acknowledge that a failure to abide by NYU Stern Honor Code will result in a failing grade for the project and course.

Project Description : 

"If you don’t have the knowledge you’re lost. A lot of people talk but they don’t have the knowledge, and knowledge is power.” - Rosa de la Cruz (former attorney for the Guggenheim Museum). 

In this day and age, we might interpret Cruz's words as alluding to an understanding of current trends in the global art market. For instance: what makes a da Vinci worth over 450 million dollars 500 years after its conception, while a work by a contemporary artist has never peaked even 150 million? Why is it than within a span of four years, the price of an artwork can drop from 1,625,000 dollars to 665,336 dollars (Jim Beam - Observation Car - Jeff Koons)? 

By studying, plotting and mapping the avaiable data from Artprice, and also the archives of prominent auction houses we can start to answer such questions. 

For my project, I want to provide this kind of close analysis. I intend to do this by looking into the structures of the top 500 contemporary artists in 2017, and of the top 1,000 artists sold in 2016 and 2017 (using the quantile function and computing the average price, for example). I also plan on comparing the different major indexes (turnover, lots sold, record price...)of the art market between 1987 and 2017 (represented in a scatter plot). Moreover, there will be a significant geographical dimension to my analysis. First by comparing the results of artworks by country of origin. Then, by comparing in which countries the major auctions are taking place and which country has the best auction houses. This will represent a major and interesting mapping plot because it allows us to track art market trends around the world.